docs: use src/README.md as source for docs
This change uses the public-facing src/readme.md as source
for https://docs.victoriametrics.com/victoriametrics/victoriametrics-datasource/.

The `src/README.md` is readme displayed in Grafana's plugins catalogue.
It is more user friendly and focused on using the plugin. While
the current root's README.md has a lot of extra info about development
and advanced installation methods that might be not relevant to majority of users.